refactor(trace): consolidate shared span utilities into log-details/utils
